+++ app-backend/backsplash-server/src/main/scala/com/rasterfoundry/backsplash/Config.scala
@@ -7,7 +7,11 @@ object Config {
 
   object parallelism {
     private val parallelismConfig = config.getConfig("parallelism")
-    val threadPoolSize = parallelismConfig.getInt("threadPoolSize")
+    val dbThreadPoolSize = parallelismConfig.getInt("dbThreadPoolSize")
+    val http4sThreadPoolSize = parallelismConfig.getInt("http4sThreadPoolSize")
+    val blazeThreadPoolSize = parallelismConfig.getInt("blazeThreadPoolSize")
+    val blazeConnectorPoolSize =
+      parallelismConfig.getInt("blazeConnectorPoolSize")
   }
 
   object server {