+++ CHANGELOG.md
@@ -29,6 +29,7 @@
 - Ignored errors from integration tests so that reports will always be written to s3 [\#4406](https://github.com/raster-foundry/raster-foundry/pull/4406)
 - Changed how the database transactor is passed to backsplash and API servers to prevent accidentally passing implicit execution contexts where they are not wanted [\#4415](https://github.com/raster-foundry/raster-foundry/pull/4415)
 - Added ability to test against several projects in gatling integration tests [\#4416](https://github.com/raster-foundry/raster-foundry/pull/4416)
+- Changed how to configure threadpools for backsplash and hikari [\#4420](https://github.com/raster-foundry/raster-foundry/pull/4420)
 
 ### Fixed